fix test and mypy
